\name{alkalinity}
\docType{data}
\alias{alkalinity}
\title{Example data file for function at}
\description{
The variables are:

\itemize{
	\item volume: Volume of acid added to the sample in ml	
	\item E: Potential measured during the titration in mV	
	\item temperature: Temperature in degrees Celsius	
	\item weight: Weight of the sample in g
	\item S: Salinity
	\item normality : Normality of the acid
	\item ETris: Potential used for the calibration of the electrode in mV
	\item pHTris: pH used for the calibration of the electrode with the TRIS buffer
	}
}
\usage{alkalinity}

\format{A data frame with 29 rows and 8 variables}

\source{
Data come from a potentiometric titration performed by Steeve Comeau.
}

\keyword{datasets}
