\encoding{latin1}
\name{buffer}
\alias{buffer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Buffer parameters of the seawater carbonate system}
\description{Returns the eleven buffer factors of the seawater carbonate system defined by Frankignoulle (1994). For each of four possible chemical perturbations -- input or output of dissolved CO2 (D), bicarbonate (B), carbonate (C), or strong acid \eqn{H_3O^+} (H) -- Frankignoulle (1994) defines a chemical buffer factor \eqn{\Phi} (pH change per unit input/output; \eqn{\Phi_H = -dpH/d|Talk|} for the H case, \eqn{\Phi_{D,B,C} = dpH/d|\Sigma CO_2|} otherwise) and a further buffer factor \eqn{\Pi} describing the accompanying pCO2 change, in \eqn{\mu}atm (\eqn{\Pi_H = -dpCO_2/d|Talk|}; \eqn{\Pi_{D,B,C} = dpCO_2/d|\Sigma CO_2|}). For the D, B, and C cases only, a third, dimensionless factor \eqn{\beta} is also returned (\eqn{\beta_{D,B,C} = dln(pCO_2)/dln|\Sigma CO_2|}). \eqn{\beta_D} (\code{BetaD}) is the classical Revelle factor, termed the ``homogeneous buffer factor'' by Frankignoulle (1994); that terminology, and the underlying quantity itself, trace back to Sundquist et al. (1979). This implementation follows Frankignoulle (1994) for the borate- and water-inclusive (``practical alkalinity'') case, extended here to also account for phosphoric and silicic acid systems when Pt and Sit are nonzero, as well as NH3 and H2S acid-base systems when NH4t and HSt are nonzero, and has since been cross-validated against seacarb's \code{buffesm} (Egleston et al., 2010) and \code{buffzwg} (Zeebe and Wolf-Gladrow, 2001, corrected): \code{BetaD} agrees with \code{buffesm}'s \code{R} and \code{buffzwg}'s \code{RF} to floating-point precision. Its input arguments are identical to those in the \code{buffesm} and \code{carb} functions of seacarb.}
\usage{
buffer(flag, var1, var2, S = 35, T = 25, Patm = 1, P = 0, Pt = 0, Sit = 0, 
  NH4t = 0, HSt = 0, k1k2 = "x", kf = "x", ks = "d", pHscale = "T", b = "u74", warn = "y", 
  eos = "eos80", long = 1e+20, lat = 1e+20)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{flag}{select the pair of variables available. The flags which can be used are: 

flag = 1      pH and CO2 given

flag = 2      CO2 and HCO3 given

flag = 3      CO2 and CO3 given

flag = 4      CO2 and ALK given

flag = 5      CO2 and DIC given

flag = 6      pH and HCO3 given

flag = 7      pH and CO3 given

flag = 8      pH and ALK given

flag = 9      pH and DIC given

flag = 10     HCO3 and CO3 given

flag = 11     HCO3 and ALK given

flag = 12     HCO3 and DIC given

flag = 13     CO3 and ALK given

flag = 14     CO3 and DIC given

flag = 15     ALK and DIC given

flag = 21     pCO2 and pH given

flag = 22     pCO2 and HCO3 given

flag = 23     pCO2 and CO3 given

flag = 24     pCO2 and ALK given

flag = 25     pCO2 and DIC given
}
	\item{var1}{enter value of the first  variable in mol/kg, except for pH and for pCO2 in \eqn{\mu}atm}
	\item{var2}{enter value of the second  variable in mol/kg, except for pH}
	\item{S}{Salinity}
	\item{T}{Temperature in degrees Celsius}
  \item{Patm}{Surface atmospheric pressure in atm}
  \item{P}{Hydrostatic pressure in bar (surface = 0)}
  \item{Pt}{Concentration of total phosphate in mol/kg; set to 0 if NA}
  \item{Sit}{Concentration of total silicate in mol/kg; set to 0 if NA}
  \item{NH4t}{Concentration of total ammonium in mol/kg; set to 0 if NA; when nonzero (or when \code{HSt} is nonzero), \code{carbfull()} is used internally instead of \code{carb()}}
  \item{HSt}{Concentration of total hydrogen sulfide in mol/kg; set to 0 if NA; when nonzero (or when \code{NH4t} is nonzero), \code{carbfull()} is used internally instead of \code{carb()}}
  \item{k1k2}{"cw" for using K1 and K2 from Cai & Wang (1998), "l" from Lueker et al. (2000), "m02" from Millero et al. (2002), "m06" from Millero et al. (2006), "m10" from Millero (2010), "mp2" from Mojica Prieto et al. (2002), "p18" from Papadimitriou et al. (2018), "r" from Roy et al. (1993), "sb21" from Shockman & Byrne (2021), "s20" from Sulpis et al. (2020), and "w14" from Waters et al. (2014). "x" is the default flag; the default value is then "l", except if T is outside the range 2 to 35oC and/or S is outside the range 19 to 43. In these cases, the default value is "w14".}
  \item{kf}{"pf" for using Kf from Perez and Fraga (1987) and "dg" for using Kf from Dickson and Riley (1979 in Dickson and Goyet, 1994). "x" is the default flag; the default value is then "pf", except if T is outside the range 9 to 33oC and/or S is outside the range 10 to 40. In these cases, the default is "dg".}
	\item{ks}{"d" for using Ks from Dickon (1990), "k" for using Ks from Khoo et al. (1977), default is "d"} 
	\item{pHscale}{choice of pH scale: "T" for the total scale, "F" for the free scale and "SWS" for using the seawater scale, default is "T" (total scale)}
 	\item{b}{Concentration of total boron. "l10" for the Lee et al. (2010) formulation or "u74" for the Uppstrom (1974) formulation, default is "u74" }
  \item{warn}{"y" to show warnings when T or S go beyond the valid range for constants; "n" to supress warnings. The default is "y".}
    \item{eos}{"teos10" to specify T and S according to Thermodynamic Equation Of Seawater - 2010 (TEOS-10); "eos80" to specify T and S according to EOS-80.}
  \item{long}{longitude of data point, used when eos parameter is "teos10" as a conversion parameter from absolute to practical salinity.}
  \item{lat}{latitude of data point, used when eos parameter is "teos10".}
}


\details{
The Lueker et al. (2000) constants for K1 and K2, the Perez and Fraga (1987) constant for Kf and the Dickson (1990) constant for Ks are recommended by Dickson et al. (2007). It is, however, critical to consider that each formulation is only valid for specific ranges of temperature and salinity:

\emph{For K1 and K2:}
\itemize{
\item Cai and Wang (1998): S ranging between 0 and 40 and T ranging between 0.2 and 30oC.

\item Lueker et al. (2000): S ranging between 19 and 43 and T ranging between 2 and 35oC.

\item Millero et al. (2002): S ranging from 34 to 37 and T ranging between -1.6 and 35oC.

\item Millero et al. (2006): S ranging between 0.1 and 50 and T ranging between 1 and 50oC.

\item Millero (2010): S ranging between 1 and 50 and T ranging between 0 and 50oC. Millero (2010) provides a K1 and K2 formulation for the seawater, total and free pH scales. Therefore, when this method is used and if P=0, K1 and K2 are computed with the formulation corresponding to the pH scale given in the flag "pHscale". 

\item Mojica Prieto et al. (2002): S ranging from 5 to 42 and T ranging between 0 and 45oC.

\item Papadimitriou et al. (2018): S ranging from 33 to 100 and T ranging between -6 to 25oC.

\item Roy et al. (1993): S ranging between 5 and 45 and T ranging between 0 and 45oC.

\item Shockman & Byrne (2021): for K2, S ranging from 19.6 to 41 and T ranging between 15 to 35oC. For K1, formulation is that of Waters et al.

\item Sulpis et al. (2020): S ranging from 30.7 to 37.6 and T ranging between -1.7 to 31.8oC.

\item Waters et al.(2014): S ranging between 1 and 50 and T ranging between 0 and 50oC. Waters (2014) provides a K1 and K2 formulation for the seawater, total and free pH scales. Therefore, when this method is used and if P=0, K1 and K2 are computed with the formulation corresponding to the pH scale given in the flag "pHscale".

}

\emph{For Kf:}
\itemize{
\item Perez and Fraga (1987): S ranging between 10 and 40 and T ranging between 9 and 33oC.

\item Dickson and Riley (1979 in Dickson and Goyet, 1994): S ranging between 0 and 45 and T ranging between 0 and 45oC.
}

\emph{For Ks:}
\itemize{
\item Dickson (1990): S ranging between 5 and 45 and T ranging between 0 and 45oC.

\item Khoo et al. (1977): S ranging between 20 and 45 and T ranging between 5 and 40oC.
}

The arguments can be given as a unique number or as vectors. If the lengths of the vectors are different, the longer vector is retained and only the first value of the other vectors is used. It is recommended to use either vectors with the same dimension or one vector for one argument and numbers for the other arguments.

\emph{On the non-carbonate buffering term DD:}

Writing \eqn{h := [H^+]}, \eqn{b := [HCO_3^-]}, and \eqn{c := [CO_3^{2-}]}, \code{PhiD}/\code{BetaD}/\code{PiD} are built from a non-carbonate buffering term \code{DD}, constructed additively, one acid-base system at a time (the same construction used in \code{buffzwg} and \code{buffesm}, there called \eqn{w} and \code{Segle} respectively):

\eqn{DD = w_{bw} + w_{Si} + w_P + w_N + w_S}

with \eqn{w_{bw} = 1 + K_w/h^2 + B_T K_B/(K_B+h)^2} (borate and water), \eqn{w_{Si} = Si_T K_{Si}(h^2+4K_{2Si}h+K_{Si}K_{2Si})/(h^2+K_{Si}h+K_{Si}K_{2Si})^2} (silicic acid, treated as the full diprotic system, using both \eqn{K_{Si}} and \eqn{K_{2Si}}), \eqn{w_P} the analogous contribution from the three-step phosphoric acid system, \eqn{w_N = NH_{4t} K_N/(K_N+h)^2} (ammonia), and \eqn{w_S = HS_t K_{HS}/(K_{HS}+h)^2} (hydrogen sulfide). \eqn{w_N} and \eqn{w_S} are automatically and exactly zero unless \code{NH4t} or \code{HSt}, respectively, are nonzero. The same \code{DD} is reused, unchanged, as \code{V} for the bicarbonate/carbonate input-output cases (\code{PhiB}/\code{BetaB}/\code{PiB}, \code{PhiC}/\code{BetaC}/\code{PiC}), and \code{PhiH}/\code{PiH} are computed from the identity \eqn{\Phi_H = 1/(h \ln(10) (D-DD))}{PhiH = 1/(h*ln(10)*(D-DD))}, which holds regardless of how many non-carbonate systems \code{DD} includes.

\emph{Pressure corrections and pH scale:}

\itemize{
\item For K0, the pressure correction term of Weiss (1974) is used.

\item For K1, K2, pK1, pK2, pK3, Kw, Kb, Khs and Ksi, the pressure correction was applied on the seawater scale. Hence, if needed, values were first transformed from the total scale to the seawater scale, the pressure correction applied as described by Millero (1995), and the value was transformed back to the required scale (T, F or SWS).

\item For Kf, the pressure correction was applied on the free scale. The formulation of Dickson and Riley (1979 in Dickson and Goyet, 1994) provides Kf on the free scale but that of Perez and Fraga (1987) provides it on the total scale. Hence, in that case, Kf was first transformed from the total scale to the free scale. With both formulations, the pressure correction was applied as described by Millero (1995), and the value was transformed back to the required scale (T, F or SWS). 

\item For Ks, the pressure correction was applied on the free scale. The pressure correction was applied as described by Millero (1995), and the value was transformed back to the required scale (T, F or SWS). 

\item For Kn, The pressure correction was applied on the seawater scale. The pressure correction was applied as described by Millero (1995), and the value was transformed back to the required scale (T, F or SWS).
}
long and lat are used as conversion parameters from absolute to practical salinity: when seawater is not of standard composition, practical salinity alone is not sufficient to compute absolute salinity and vice-versa. One needs to know the density. When long and lat are given, density is inferred from WOA silicate concentration at given location. When they are not, an arbitrary geographic point is chosen: mid equatorial Atlantic. Note that this implies an error on computed salinity up to 0.02 g/kg.

All eleven factors follow one convention: each is the change in pH (Phi), in
ln(fCO2) (Beta), or in fCO2 (Pi), \emph{per mole of the named substance added}.
The four substances are dissolved CO2 (D), bicarbonate (B), carbonate (C) and
strong acid (H), with

\tabular{lcc}{
  substance      \tab \eqn{\Delta}DIC \tab \eqn{\Delta}ALK \cr
  dissolved CO2  \tab  1 \tab  0 \cr
  bicarbonate    \tab  1 \tab  1 \cr
  carbonate      \tab  1 \tab  2 \cr
  strong acid    \tab  0 \tab -1 \cr
}

Two consequences are easy to misread and are stated explicitly:

\enumerate{
  \item PhiH and PiH are per mole of \emph{strong acid} added, not per mole of
        alkalinity. Since adding acid removes alkalinity, PhiH = -dpH/d[ALK].
        PhiH is therefore negative even though dpH/d[ALK] is positive.
  \item The Beta and Pi factors are computed from \emph{fCO2}, not pCO2. The two
        differ by about 0.3 percent at 20 degrees C. The Pi factors are returned
        per \eqn{\mu}mol/kg, not per mol/kg.
}

Every factor has been verified against finite differences taken through
\code{carb()}, and agrees to better than 1e-6 percent.
}

\value{The function returns a data frame containing the following columns:

  \item{PhiD}{PhiD, chemical buffer factor for input/output of dissolved CO2: the change in pH per mole of CO2 added (\eqn{\Delta}DIC = 1, \eqn{\Delta}ALK = 0). Units: pH per mol/kg.}
  \item{BetaD}{BetaD, homogeneous or Revelle buffer factor for input/output of dissolved CO2: dln(fCO2)/dln(DIC) at constant ALK. Dimensionless.}
  \item{PiD}{PiD, chemical buffer factor for input/output of dissolved CO2: the change in fCO2 per mole of CO2 added. Units: \eqn{\mu}atm per \eqn{\mu}mol/kg.}
  \item{PhiB}{PhiB, chemical buffer factor for input/output of bicarbonate: the change in pH per mole of HCO3- added (\eqn{\Delta}DIC = 1, \eqn{\Delta}ALK = 1). Units: pH per mol/kg.}
  \item{BetaB}{BetaB, homogeneous buffer factor for input/output of bicarbonate: dln(fCO2)/dln(DIC) along that perturbation. Dimensionless.}
  \item{PiB}{PiB, chemical buffer factor for input/output of bicarbonate: the change in fCO2 per mole of HCO3- added. Units: \eqn{\mu}atm per \eqn{\mu}mol/kg.}
  \item{PhiC}{PhiC, chemical buffer factor for input/output of carbonate: the change in pH per mole of CO3-- added (\eqn{\Delta}DIC = 1, \eqn{\Delta}ALK = 2). Units: pH per mol/kg.}
  \item{BetaC}{BetaC, homogeneous buffer factor for input/output of carbonate: dln(fCO2)/dln(DIC) along that perturbation. Dimensionless.}
  \item{PiC}{PiC, chemical buffer factor for input/output of carbonate: the change in fCO2 per mole of CO3-- added. Units: \eqn{\mu}atm per \eqn{\mu}mol/kg.}
  \item{PhiH}{PhiH, chemical buffer factor for input/output of strong acid: the change in pH per mole of strong acid added (\eqn{\Delta}ALK = -1, \eqn{\Delta}DIC = 0). Hence PhiH = -dpH/d[ALK], and is NEGATIVE: adding acid lowers pH. Units: pH per mol/kg.}
  \item{PiH}{PiH, chemical buffer factor for input/output of strong acid: the change in fCO2 per mole of strong acid added. Units: \eqn{\mu}atm per \eqn{\mu}mol/kg.}
}

\note{When \code{Sit > 0} and \code{NH4t = HSt = 0}, this function's diprotic treatment of silicate alkalinity (using both \code{Ksi} and \code{K2si}) is very slightly inconsistent with the underlying carbonate system state, because that state comes from \code{carb()}, which internally treats silicate as monoprotic (\code{K2si = 0}) regardless of \code{Sit} (see \code{carb.Rd}). \code{carbfull()} is only invoked here when \code{NH4t} or \code{HSt} are nonzero, so this inconsistency is not corrected for when only \code{Pt}/\code{Sit} are supplied; it is expected to affect results at only the ~1e-6 relative level. See \code{carb.Rd} and \code{carbfull.Rd} for further discussion.

Frankignoulle's (1994) own printed equation for \eqn{\Pi_H} (his Eq. 11) contains an error: as published, \eqn{\Pi_H} is given as a quotient divided by \eqn{\Phi_H}, whereas it should be their product. This was corrected in seacarb's implementation by J.-P. Gattuso, working with M. Frankignoulle. Later, we independently reconfirmed the fix here by rederiving \eqn{\Pi_H} from scratch and checking it against the \eqn{\Pi_H} values listed in Frankignoulle's (1994) Table 1. Rederiving each of the other buffer factor equations in the same way turned up no further errors, including for \eqn{\beta_D} (\code{BetaD}, the Revelle factor), which matches Frankignoulle's (1994) original formulation exactly.}




\references{
Cai W. J., and Wang Y., 1998. The chemistry, fluxes, and sources of carbon dioxide in the estuarine waters of the Satilla and Altamaha Rivers, Georgia. \emph{Limnology and Oceanography} \bold{43}, 657-668.

Dickson A. G., 1990 Standard potential of the reaction: AgCI(s) + 1/2H2(g) = Ag(s) + HCI(aq), and the standard acidity constant of the ion HSO4 in synthetic sea water from 273.15 to 318.15 K. \emph{Journal of Chemical Thermodynamics} \bold{22}, 113-127.

Dickson A. G., Sabine C. L. and Christian J. R., 2007 Guide to best practices for ocean CO2 measurements. \emph{PICES Special Publication} \bold{3}, 1-191.

Khoo H. K., Ramette R. W., Culberson C. H. and Bates R. G., 1977 Determination of Hydrogen Ion Concentration in Seawater from 5 to 40oC: Standard Potentials at Salinities from 20 to 45. \emph{Analytical Chemistry} \bold{49}, 29-34. 

Frankignoulle M., 1994 A complete set of buffer factors for acid/base CO2 system in seawater. \emph{Journal of Marine Systems} \bold{5}, 111-118.

Lee K., Tae-Wook K., Byrne R.H., Millero F.J., Feely R.A. and Liu Y-M, 2010 The universal ratio of the boron to chlorinity for the North Pacific and North Atlantoc oceans. \emph{Geochimica et Cosmochimica Acta} \bold{74} 1801-1811.

Lueker T. J., Dickson A. G. and Keeling C. D., 2000 Ocean pCO2 calculated from dissolved inorganic carbon, alkalinity, and equations for K1 and K2: validation based on laboratory measurements of CO2 in gas and seawater at equilibrium. \emph{Marine Chemistry} \bold{70} 105-119.

Millero F. J., 2010 Carbonate constant for estuarine waters. \emph{Marine and Freshwater Research} \bold{61}: 139-142.

Millero F. J., Graham T. B., Huang F., Bustos-Serrano H. and Pierrot D., 2006 Dissociation constants of carbonic acid in seawateras a function of salinity and temperature.  \emph{Marine Chemistry} \bold{100}, 80-84.  

Perez F. F. and Fraga F., 1987 Association constant of fluoride and hydrogen ions in seawater. \emph{Marine Chemistry} \bold{21}, 161-168.

Roy R. N., Roy L. N., Vogel K. M., Porter-Moore C., Pearson T., Good C. E., Millero F. J. and Campbell D. M., 1993. The dissociation constants of carbonic acid in seawater at salinities 5 to 45 and temperatures 0 to 45oC. \emph{Marine Chemistry} \bold{44}, 249-267.

Schockman, K.M., Byrne, R.H., 2021. Spectrophotometric determination of the bicarbonate dissociation constant in seawater, \emph{Geochimica et Cosmochimica Acta}.

Sundquist E. T., Plummer L. N. and Wigley T. M. L., 1979 Carbon dioxide in the ocean surface: the homogeneous buffer factor. \emph{Science} \bold{204}, 1203-1205.

Uppstrom L.R., 1974 The boron/chlorinity ratio of the deep-sea water from the Pacific Ocean. \emph{Deep-Sea Research I} \bold{21} 161-162.
}

\author{
Heloise Lavigne, Aurelien Proye and Jean-Pierre Gattuso \email{jean-pierre.gattuso@imev-mer.fr}
}

\examples{

## Computation with a pair of variables
buffer(flag=8, var1=8.2, var2=0.00234, S=35, T=25, Patm=1, P=0, Pt=0, 
	Sit=0, pHscale="T", kf="pf", k1k2="l", b="u74")

## Using vectors as arguments
flag <- c(8, 2, 8)
var1 <- c(8.2, 7.477544e-06, 8.2)
var2 <- c(0.002343955, 0.001649802, 2400e-6)
S <- c(35, 35, 30)
T <- c(25, 25, 30)
P <- c(0, 0, 0)
Pt <- c(0, 0, 0)
Sit <- c(0, 0, 0)
kf <- c("pf", "pf", "pf")
k1k2 <- c("l", "l", "l")
pHscale <- c("T", "T", "T")
b <- c("l10", "l10", "l10")
buffer(flag=flag, var1=var1, var2=var2, S=S, T=T, P=P, Pt=Pt, 
	Sit=Sit, kf=kf, k1k2=k1k2, pHscale=pHscale, b=b)


## Test for all flags; all rows describe the SAME equilibrium seawater
## state expressed 20 different ways, so results should be (almost)
## identical across rows -- tiny differences reflect solver precision,
## not a real difference in the underlying state.

flag <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 21, 22, 23, 24, 25)

var1 <- c(8.200000, 7.477544e-06, 7.477544e-06, 7.477544e-06, 7.477544e-06, 8.2, 
	8.2, 8.2, 8.2, 0.001685024, 0.001685024, 0.001685024,  0.0002888382, 
	0.0002888382, 0.002391252, 264.2008, 264.2008, 264.2008, 264.2008, 264.2008)

var2 <- c(7.477544e-06, 0.001685024, 0.0002888382, 0.002391252, 0.001981340, 
	0.001685024, 0.0002888382, 0.002391252, 0.001981340, 0.0002888382, 0.002391252,
	0.001981340,  0.002391252, 0.001981340, 0.001981340, 8.2, 0.001685024, 
	0.0002888382, 0.002391252, 0.001981340)

buffer(flag=flag, var1=var1, var2=var2)

## BetaD is the Revelle factor (Frankignoulle's "homogeneous buffer factor"):
## cross-check against seacarb's buffesm() R and buffzwg() RF, which should
## agree to floating-point precision
bu <- buffer(flag=15, var1=0.00230, var2=0.00200, S=35, T=18)
es <- buffesm(flag=15, var1=0.00230, var2=0.00200, S=35, T=18)
zw <- buffzwg(flag=15, var1=0.00230, var2=0.00200, S=35, T=18)
bu$BetaD - es$R    # should be ~0 (floating-point precision)
bu$BetaD - zw$RF   # should be ~0 (floating-point precision)

## Effect of adding ammonia and hydrogen sulfide alkalinity (NH4t, HSt);
## when either is nonzero, carbfull() is used internally instead of carb()
buffer(flag=8, var1=8.2, var2=0.00234, S=35, T=25, Pt=0, Sit=0, NH4t=0, HSt=0)
buffer(flag=8, var1=8.2, var2=0.00234, S=35, T=25, Pt=0, Sit=0, NH4t=10e-6, HSt=1e-6)

}
\keyword{utilities}
