\name{seaFET}

\docType{data}

\alias{seaFET}

\title{Test seaFET data file}

\description{
Short test file for using with functions sf_calc and sf_calib. It is an excerpt of a file produced by a SeaFET pH sensor.
}
\usage{seaFET}

\format{
A data frame with 8 variables (datetime, Eint, Eext, Salinity, Temperature, pHspectro, E0int25, E0ext25) and 10 rows
}

\keyword{datasets}
